ui(operations): extract EmbeddingTab into its own module
Lifts the embedding tab body (stat row, provider banner, namespace
breakdown table, action panels, confirm dialog wiring) out of
GroveMaintenance.tsx so the upcoming Operations page shell can
compose it. Behavior is unchanged: same hooks, same handlers, same
scope-pinned-to-grove logic.

Also drops a dead BACKUP_SCOPE_AVAILABLE constant the prior version
declared but never referenced.
